150920539 Add more documentation

+++ server/controllers/favouriteRecipe.js
@@ -1,32 +1,33 @@
 
 import allModels from '../models';
+import MiddleWare from '../middleware';
+
+const middleware = new MiddleWare();
 
 const favouriteRecipeModel = allModels.favourite_recipes;
 const recipeModel = allModels.recipes;
 /**
  * get all favourites recipes of a user
- * @param  {} req
- * @param  {} res
+ * @param  {object} req request object
+ * @param  {object} res response object
+ * @returns {null} does not return a value
  */
 const getAllFavoriteRecipes = (req, res) => {
-  favouriteRecipeModel.findAll(
-    {
-      where: {
-        userId: req.params.userId
-      },
-      include: [
-        { model: recipeModel }
-      ],
-    }
-  ).then(output => res.json({ status: 'success', data: output })).catch(error => res.json(error.toString()));
+  favouriteRecipeModel.findAll({
+    where: {
+      userId: req.params.userId
+    },
+    include: [
+      { model: recipeModel }
+    ],
+  }).then(output => res.json({ status: 'success', data: output })).catch(error => res.json(error.toString()));
 };
-/**
- * Add a recipe to favourites list of a user
- * @param  {} req
- * @param  {} res
- */
 const addFavoriteRecipe = (req, res) => {
   if (req.params.recipeId) {
+    if (!middleware.validateStringIsNumber(req.params.recipeId)) {
+      res.json({ status: 'fail', message: 'recipe id must be a number' });
+      return;
+    }
     favouriteRecipeModel.findAndCountAll({
       where: {
         userId: req.decoded.id,
@@ -36,14 +37,16 @@ const addFavoriteRecipe = (req, res) => {
       if (output.count === 1) {
         res.json({ status: 'fail', message: 'This recipe already exist on your list of favourite recipes' });
       } else if (output.count === 0) {
-        favouriteRecipeModel.create(
-          {
-            userId: req.decoded.id,
-            recipeId: req.params.recipeId
-          }
-        ).then((output) => {
+        favouriteRecipeModel.create({
+          userId: req.decoded.id,
+          recipeId: req.params.recipeId
+        }).then((output) => {
           recipeModel.findById(req.params.recipeId).then((response) => {
-            res.json({ status: 'success', data: { favourite: output, recipe: response } });
+            res.json({
+              status: 'success',
+              message: 'Recipe successfully added to your favourites list',
+              data: { favourite: output, recipe: response }
+            });
           }).catch(error => res.send(error));
         }).catch(error => res.send(error));
       }
@@ -54,19 +57,26 @@ const addFavoriteRecipe = (req, res) => {
 };
 /**
  * Remove recipe from favourite recipes list
- * @param  {} req
- * @param  {} res
+ * @param  {object} req request object
+ * @param  {object} res response object
+ * @returns {null} returns null
  */
 const removeFromFavoriteRecipes = (req, res) => {
-  favouriteRecipeModel.destroy(
-    {
-      where: {
-        userId: req.decoded.id,
-        recipeId: req.params.recipeId
-      }
+  if (!middleware.validateStringIsNumber(req.params.recipeId)) {
+    res.json({ status: 'fail', message: 'recipe id must be a number' });
+    return;
+  }
+  favouriteRecipeModel.destroy({
+    where: {
+      userId: req.decoded.id,
+      recipeId: req.params.recipeId
     }
-  ).then(output => res.json({ status: 'success', message: 'Recipe removed successfully' })).catch(error => res.send(error));
+  }).then(output => res.json({ status: 'success', message: 'Recipe removed successfully' })).catch(error => res.send(error));
 };
 
-const allMethods = { getFavouriteRecipes: getAllFavoriteRecipes, addFavourite: addFavoriteRecipe, removeFavourite: removeFromFavoriteRecipes };
+const allMethods = {
+  getFavouriteRecipes: getAllFavoriteRecipes,
+  addFavourite: addFavoriteRecipe,
+  removeFavourite: removeFromFavoriteRecipes
+};
 export default allMethods;
